[D3D12] Added physical device and display mode detection. [RHI] Moved physical device and display mode detection functions to RHI_Device since they are required by every API (not something that should be hidden away and can possibbly not implemented).

+++ Runtime/RHI/RHI_CommandList.h
@@ -177,7 +177,7 @@ namespace Spartan
         // Variables to minimise state changes
         uint32_t m_vertex_buffer_id     = 0;
         uint64_t m_vertex_buffer_offset = 0;
-        uint32_t m_index_buffer_id      = 0;
+        uint64_t m_index_buffer_id      = 0;
         uint64_t m_index_buffer_offset  = 0;
     };
 }
